requirements.txt
streamlit
crewai
langchain
langchain-openai
langchain-anthropic
langchain-mistralai
python-dotenv
